requirements.txt
PyYaml==6.0.2
tqdm==4.67.1
requests==2.32.3
loguru==0.7.3
spacy==3.8.5
pandas==2.2.3
Levenshtein==0.26.1
rstr==3.2.2
prance==23.6.21.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.6.0
openapi3-parser==1.1.19
nltk==3.9.1
click==8.1.8
seaborn==0.13.2
matplotlib==3.9.2
openai==2.0.0
scikit-learn==1.6.1
tiktoken==0.5.2
httpx==0.27.2